Computing The Greeks

Background

The Greeks is a term used in the options market to describe the different dimensions of risk involved in taking an options position. These variables are called Greeks because they are typically associated with Greek symbols. Each risk variable is a result of an imperfect assumption or relationship of the option with another underlying variable. Traders use different Greek values, such as delta, theta, and others, to assess options risk and manage option portfolios.

PooksoftOptionsKit.delta — Function

delta(assetSet::Set{PSAbstractAsset}, parameters::PSBinaryLatticeModel, underlyingAssetPrice::Float64; 
    earlyExercise::Bool = false)::PSResult

Compute the change in the option price for a 1-dollar increase in the underlying stock price

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call models involved in this trade.

parameters::PSBinaryLatticeModel: A PSBinaryLatticeModel object containing the parameters for the lattice

underlyingAssetPrice::Float64: Underlying stock price at the time of purchase of the contract

earlyExercise::Bool = false: Can this option contract be excercised early (true for American options)

PooksoftOptionsKit.theta — Function

theta(assetSet::Set{PSAbstractAsset}, parameters::PSBinaryLatticeModel, underlyingAssetPrice::Float64; 
    earlyExercise::Bool = false)::PSResult

Compute the change in the price of the option contract for a one day decrease in the number of days left until expiration

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call models involved in this trade.

parameters::PSBinaryLatticeModel: A PSBinaryLatticeModel object containing the parameters for the lattice

underlyingAssetPrice::Float64: Underlying stock price at the time of purchase of the contract

earlyExercise::Bool = false: Can this option contract be excercised early (true for American options)

PooksoftOptionsKit.gamma — Function

gamma(assetSet::Set{PSAbstractAsset}, parameters::PSBinaryLatticeModel, underlyingAssetPrice::Float64; 
    earlyExercise::Bool = false)::PSResult

Compute the rate of change of the delta parameter

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call models involved in this trade.

parameters::PSBinaryLatticeModel: A PSBinaryLatticeModel object containing the parameters for the lattice

underlyingAssetPrice::Float64: Underlying stock price at the time of purchase of the contract

earlyExercise::Bool = false: Can this option contract be excercised early (true for American options)

PooksoftOptionsKit.vega — Function

vega(assetSet::Set{PSAbstractAsset}, parameters::PSBinaryLatticeModel, underlyingAssetPrice::Float64; 
    earlyExercise::Bool = false)::PSResult

Compute the change in the price of an option contract for a 1-percent increase in the implied volatility

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call models involved in this trade.

parameters::PSBinaryLatticeModel: A PSBinaryLatticeModel object containing the parameters for the lattice

underlyingAssetPrice::Float64: Underlying stock price at the time of purchase of the contract

earlyExercise::Bool = false: Can this option contract be excercised early (true for American options)

PooksoftOptionsKit.rho — Function

rho(assetSet::Set{PSAbstractAsset}, parameters::PSBinaryLatticeModel, underlyingAssetPrice::Float64; 
    earlyExercise::Bool = false)::PSResult

Compute the change in the price of an option contract for a 1% increase in the risk free rate

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call models involved in this trade.

parameters::PSBinaryLatticeModel: A PSBinaryLatticeModel object containing the parameters for the lattice

underlyingAssetPrice::Float64: Underlying stock price at the time of purchase of the contract

earlyExercise::Bool = false: Can this option contract be excercised early (true for American options)