Profit and loss simulations at expiration

Background

Some background stuff goes here.

PooksoftOptionsKit.compute_option_profit_and_loss_at_expiration — Function

compute_option_profit_and_loss_at_expiration(assetSet::Set{PSAbstractAsset}, assetPriceArray::Array{Float64,1})::(Union{PooksoftBase.PSResult{T}, Nothing} where T<:Any)

Compute the overall profit and loss (P/L) for a set option contracts with the same expiration date, and underlying asset.

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call contract models in this trade.

assetPriceArray::Array{Float64,1}: An 1d array containing underlying asset prices to be used in the P/L calculation for this trade

compute_option_profit_and_loss_at_expiration(assetSet::Set{PSAbstractAsset}, assetPriceStart::Float64, assetPriceStop::Float64; 
    number_of_price_steps::Int64=1000)::(Union{PooksoftBase.PSResult{T}, Nothing} where T<:Any)

Compute the overall profit and loss (P/L) for a set option contracts with the same expiration date, and underlying asset.

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call contract models in this trade.

assetPriceStart::Float64: The start price for the underlying asset used to calculate the P/L values in this trade

assetPriceStop::Float64: The start price for the underlying asset used to calculate the P/L values in this trade

number_of_price_steps::Int64=1000: keyword arg describing the number of steps to take between the start and stop price. Default: 1000

compute_option_profit_and_loss_at_expiration(assetSet::Set{PSAbstractAsset}, underlyingPriceRange::Tuple{Float64,Float64,Int64})::(Union{PooksoftBase.PSResult{T}, Nothing} where T<:Any)

Compute the overall profit and loss (P/L) for a set option contracts with the same expiration date, and underlying asset.

Arguments

assetSet::Set{PSAbstractAsset}: A set containing the put and call contract models in this trade.

underlyingPriceRange::Tuple{Float64,Float64,Int64}: A tuple containing the price start, price stop and number of steps between the start and stop price to be used in the P/L calculation for this trade